requirements.txt
robotframework==4.0.3
selenium==4.9.0
pillow==9.5.0
opencv-python==4.7.0.72
numpy==1.23.4
pytesseract==0.3.9
requests==2.28.2
